fix: css.ts files can't export functions
