+++ lib/components/Box/boxStyles.ts
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import type { ElementType } from 'react';
 
-import { resetStyles } from '../../styles/reset.css';
+import { resetStyles } from '../../styles/resetStyles';
 import {
 	sprinkles,
 	sprinklesLegacyColours,
